pyproject.toml
[tool.poetry]
name = "novabot2"
version = "0.0.2"
description = "Yet another nonebot2 bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.5"

[tool.nonebot]
plugins = []
plugin_dirs = ["novabot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
